# backend/src/docs/api.yaml
...
openapi: 3.0.0
info:
  title: Airport Passenger Rights API
  description: |
    # Airport Passenger Rights Platform API
    
    This API powers the Airport Passenger Rights platform, providing:
    - Baggage allowance calculations
    - Passenger rights information
    - Complaint letter generation
    - Airline rules management
    
    ## Base URL
    - Production: `https://api.passengerrights.com/v1`
    - Staging: `https://staging-api.passengerrights.com/v1`
    - Development: `http://localhost:3001/api`
    
    ## Authentication
    - Public endpoints: No authentication required
    - Admin endpoints: JWT Bearer token required
    
    ## Rate Limits
    - Public: 100 requests per 15 minutes per IP
    - Authenticated: 1000 requests per 15 minutes per user
  version: 1.0.0
  contact:
    name: API Support
    email: api@passengerrights.com
    url: https://passengerrights.com/contact
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.passengerrights.com/v1
    description: Production server
  - url: https://staging-api.passengerrights.com/v1
    description: Staging server
  - url: http://localhost:3001/api
    description: Development server
...
paths:
...
  /baggage/calculate:
    post:
      tags: [Baggage]
      summary: Calculate baggage allowance and excess fees
      description: |
        Calculate baggage allowance based on airline, route, cabin class, and actual baggage.
        Returns allowance details, excess calculations, fees, and status indicators.
      operationId: calculateBaggage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaggageCalculationRequest'
            examples:
              domesticEconomy:
                summary: Domestic Economy Flight
                value:
                  airlineCode: "AI"
                  routeType: "DOMESTIC"
                  cabinClass: "ECONOMY"
                  passengerType: "ADULT"
                  cabinBags: [{ weight: 8 }]
                  checkedBags: [{ weight: 20 }, { weight: 15 }]
              internationalBusiness:
                summary: International Business Class
                value:
                  airlineCode: "6E"
                  routeType: "INTERNATIONAL"
                  cabinClass: "BUSINESS"
                  passengerType: "ADULT"
                  cabinBags: [{ weight: 12 }]
                  checkedBags: [{ weight: 32 }]
      responses:
        '200':
          description: Successful calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaggageCalculationResponse'
        '400':
          description: Invalid input parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Rules not found for given criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
        '500':
          description: Internal server error
  /baggage/rules/{airlineCode}:
    get:
      tags: [Baggage]
      summary: Get baggage rules for an airline
      description: Retrieve all baggage rules for a specific airline, optionally filtered by route type and cabin class.
      operationId: getAirlineBaggageRules
      parameters:
        - name: airlineCode
          in: path
          required: true
          schema:
            type: string
            example: "AI"
          description: IATA airline code (2 characters)
        - name: routeType
          in: query
          required: false
          schema:
            type: string
            enum: [DOMESTIC, INTERNATIONAL]
          description: Filter by route type
        - name: cabinClass
          in: query
          required: false
          schema:
            type: string
            enum: [ECONOMY, PREMIUM_ECONOMY, BUSINESS, FIRST]
          description: Filter by cabin class
        - name: passengerType
          in: query
          required: false
          schema:
            type: string
            enum: [ADULT, CHILD, INFANT]
            default: ADULT
          description: Passenger type
      responses:
        '200':
          description: List of baggage rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BaggageRule'
        '404':
          description: Airline or rules not found
        '500':
          description: Internal server error

  # Rights Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BaggageCalculationRequest:
      type: object
      required:
        - airlineCode
        - routeType
        - cabinClass
      properties:
        airlineCode:
          type: string
          description: IATA airline code (2 characters)
          example: "AI"
          minLength: 2
          maxLength: 2
        routeType:
          type: string
          enum: [DOMESTIC, INTERNATIONAL]
          description: Type of flight route
          example: "DOMESTIC"
        cabinClass:
          type: string
          enum: [ECONOMY, PREMIUM_ECONOMY, BUSINESS, FIRST]
          description: Cabin class
          example: "ECONOMY"
        passengerType:
          type: string
          enum: [ADULT, CHILD, INFANT]
          default: ADULT
          description: Type of passenger
        cabinBags:
          type: array
          description: List of cabin bags with weights
          items:
            type: object
            properties:
              weight:
                type: number
                minimum: 0
                description: Weight in kilograms
                example: 7.5
          default: []
        checkedBags:
          type: array
          description: List of checked bags with weights
          items:
            type: object
            properties:
              weight:
                type: number
                minimum: 0
                description: Weight in kilograms
                example: 20
          default: []
...
    BaggageCalculationResponse:
      type: object
      properties:
        airline:
          type: string
          example: "AI"
        routeType:
          type: string
          example: "DOMESTIC"
        cabinClass:
          type: string
          example: "ECONOMY"
        allowances:
          type: object
          properties:
            cabin:
              $ref: '#/components/schemas/BaggageAllowance'
            checked:
              $ref: '#/components/schemas/BaggageAllowance'
        yourBags:
          type: object
          properties:
            cabin:
              $ref: '#/components/schemas/BaggageSummary'
            checked:
              $ref: '#/components/schemas/BaggageSummary'
        calculations:
          type: object
          properties:
            cabinExcess:
              type: number
              description: Excess cabin baggage weight in kg
              example: 2.5
            checkedExcess:
              type: number
              description: Excess checked baggage weight in kg
              example: 5
            totalExcess:
              type: number
              description: Total excess weight in kg
              example: 7.5
            excessFee:
              type: number
              description: Calculated excess fee
              example: 3750
            currency:
              type: string
              example: "INR"
        status:
          type: object
          properties:
            cabin:
              type: string
              enum: [ALLOWED, EXTRA_FEE, NOT_ALLOWED]
              example: "EXTRA_FEE"
            checked:
              type: string
              enum: [ALLOWED, EXTRA_FEE, NOT_ALLOWED]
              example: "ALLOWED"
            overall:
              type: string
              enum: [ALLOWED, EXTRA_FEE, NOT_ALLOWED]
              example: "EXTRA_FEE"
        notes:
          type: string
          example: "Online prepayment available for excess baggage"
        policyUrl:
          type: string
          format: uri
          example: "https://www.airindia.com/baggage"
        disclaimer:
          type: string
          example: "Estimates only. Actual fees may vary at airport."
    BaggageAllowance:
      type: object
      properties:
        count:
          type: integer
          example: 1
        weight:
          type: number
          example: 7
        dimensions:
          type: string
          example: "55x40x20"
    BaggageSummary:
      type: object
      properties:
        count:
          type: integer
          example: 2
        totalWeight:
          type: number
          example: 15.5
        bags:
          type: array
          items:
            type: object
            properties:
              weight:
                type: number
                example: 8
...
    BaggageRule:
      type: object
      properties:
        id:
          type: integer
          example: 1
        airline_id:
          type: integer
          example: 1
        airline_name:
          type: string
          example: "Air India"
        iata_code:
          type: string
          example: "AI"
        cabin_class:
          type: string
          example: "ECONOMY"
        route_type:
          type: string
          example: "DOMESTIC"
        passenger_type:
          type: string
          example: "ADULT"
        cabin_baggage_count:
          type: integer
          example: 1
        cabin_baggage_weight:
          type: number
          example: 7
        cabin_baggage_dimensions:
          type: string
          example: "55x40x20"
        checked_baggage_count:
          type: integer
          example: 1
        checked_baggage_weight:
          type: number
          example: 15
        checked_baggage_size:
          type: string
          example: "158cm"
        excess_fee_per_kg:
          type: number
          example: 500
        excess_fee_flat:
          type: number
          example: 1000
        currency:
          type: string
          example: "INR"
        effective_from:
          type: string
          format: date
          example: "2024-01-01"
        effective_to:
          type: string
          format: date
          nullable: true
          example: "2024-12-31"
        is_active:
          type: boolean
          example: true
        notes:
          type: string
          example: "Online check-in baggage allowance may differ"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Validation Error"
        message:
          type: string
          example: "Invalid airline code provided"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "airlineCode"
              message:
                type: string
                example: "Must be a valid IATA code"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          example: "req_123456"